uploaded preliminary new design
layout and schematic likely can be cleaned up and refined. some of the traces may be cursed. need to find and assign jlcpcb components to each new part. maybe can move microSD to top of board instead. must verify design and make sure everything is connected as stuff was moved around a bit